Fix dynamic properties by making sure context hook is pulled from a nested component
